+++ Sprint_3/code/backend/src/validations/user.validation.js
@@ -2,12 +2,18 @@ const { z } = require("zod");
 const { Role } = require("@prisma/client");
 
 //Contributer: Nattawadee Chaleechat [Description] แก้ไข password validation เพิ่มเงื่อนไขต้องมีอย่างน้อย 8 ตัว ห้ามมี spaces และมีคำอย่างน้อย 3 คำ
-const { isThreeRealWords } = require("../validations/passwordValidator");
+const {
+  isThreeRealWords,
+  isBlacklistedPassword,
+} = require("../validations/passwordValidator");;
 
 const passwordRules = z
   .string()
   .min(8, "Password must be at least 8 characters")
   .regex(/^\S+$/, "Password must not contain spaces")
+  .refine((val) => !isBlacklistedPassword(val), {
+    message: "Password is too common and not allowed",
+  })
   .refine(isThreeRealWords, {
     message:
       "Password must contain at least 3 real words (e.g. apple-mango-banana)",
@@ -15,8 +21,7 @@ const passwordRules = z
 
 const createUserSchema = z.object({
   email: z.string().email("Invalid email format"),
-  username: z.string().min(6, "username is require"),
-  //Contributer: Nattawadee Chaleechat [Description] แก้ไข password validation
+  username: z.string().min(6, "username is require"),  //Contributer: Nattawadee Chaleechat [Description] แก้ไข password validation
   password: passwordRules,
   firstName: z.string().min(1, "firstname is require"),
   lastName: z.string().min(1, "lastname is require"),